% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_as.R
\name{get_as}
\alias{get_as}
\title{Pathway Feature Set Construction}
\usage{
get_as(sgi_params, sheet_name, cluster_number, cluster_names)
}
\arguments{
\item{annotations}{A dataframe with the first column consisting on the metabolite names and the second column consisting of the corresponding annotations}
}
\value{
Returns a list of lists with a set of metabolite names in each list
}
\description{
This is a function which constructs feature sets for AutoSGI based on biomolecular pathway annotations
}
\examples{
\dontrun{
}


}
