% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_selection.R
\name{hierarchical_selection}
\alias{hierarchical_selection}
\title{Hierarchical selection}
\usage{
hierarchical_selection(
  rule = rule_init(),
  sgi_params,
  cluster_min = NULL,
  cluster_max = NULL,
  metabolite_dist = "euclidean",
  metabolite_linkage = "ward.D2",
  correction_opt = "extended_simes",
  plot = T,
  supp_plot = F,
  summary_plot = F,
  return_dend = F,
  output_names = list(sgi_plots = "hierarchical-selection-results.pdf", summary =
    "hierarchical-selection-summary.pdf", cluster_results =
    "hierarchical-selection-labels.xlsx", cluster_indices =
    "hierarchical-selection-cluster-indices.xlsx", sgi_as_results =
    "hierarchical-selection-association-results.xlsx")
)
}
\arguments{
\item{rule}{A \code{rule} object containing a target clinical outcome, maximum level, and p value}

\item{sgi_params}{A \code{sgi_params} object containing information including the data, clinical outcomes, etc}

\item{cluster_min}{The minimum cluster size that should be considered for input to SGI (at least 2)}

\item{cluster_max}{The maximum cluster size that should be considered for input to SGI}

\item{metabolite_dist}{Distance metric for clustering metabolites}

\item{metabolite_linkage}{Linkage method for metabolite hierarchical clustering}

\item{correction_opt}{The method used for multiple testing corrected, either simes or bonferroni}

\item{plot}{Whether to display plots of sgi trees that are of interest according to the selection rule. Since there are many subclusters, it is recommended to do this individually after getting the results from the function.}

\item{supp_plot}{Whether to display supplementary plots to the sgi trees -- this includes the location of the cluster in the dendrogram and a list of biomolecules if a cluster size is small}

\item{return_dend}{Whether to return the dendrogram generated by hierarchically clustering the dataset}

\item{output_names}{The output names of the AutoSGI files}
}
\value{
Always returns a list of all metabolite clusters, as well as a boolean vector indicating if each tree generated by the subcluster is of interest based on the \code{rule} object. Optionally returns the tree with metabolite clusters as nodes.
}
\description{
This is a function which automatically generates metabolite clusters using hierarchical clustering and find all trees of interest from each subcluster.
}
\examples{
\dontrun{
}


}
