% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_fsets.R
\name{pathway_fsets}
\alias{pathway_fsets}
\title{Pathway Feature Set Construction}
\usage{
pathway_fsets(annotations)
}
\arguments{
\item{annotations}{A dataframe with the first column consisting on the metabolite names and the second column consisting of the corresponding annotations}
}
\value{
Returns a list of lists with a set of metabolite names in each list
}
\description{
This is a function which constructs feature sets for AutoSGI based on biomolecular pathway annotations
}
\examples{
\dontrun{
}


}
