% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Correlation outcome plotting}
\usage{
plot_correlation(clins)
}
\arguments{
\item{dataset}{A dataset with patient rows and clinical outcome columns}
}
\value{
A correlation matrix of the clinical data based on Kendall correlation
}
\description{
Correlation outcome plotting
}
