% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-rule-object.R
\name{print.rule}
\alias{print.rule}
\title{Print \code{rule} object summary}
\usage{
\method{print}{rule}(rule)
}
\arguments{
\item{rule}{an object of the class \code{rule}}
}
\value{
a description of the clinical outcome, max level, and p value contained in the object
}
\description{
This function prints a summary of all the variables contained within a rule object
}
\examples{
\dontrun{
rule1 <- rule_init("BMI", 3, 0.05)
rule2 <- rule_init("BIM", 3, 0.1)
rule3 <- rule_init("Cholesterol", 3, -1) #invalid
rule4 <- rule_init("Cholesterol", 3, 1.00001) #invalid
rule5 <- rule_init("Diabetes", 2, 0.05)
print(rule1)
print(rule2)
print(rule3)
print(rule4)
print(rule5)
}
}
