% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-rule-object.R
\name{rule_init}
\alias{rule_init}
\title{Initialize \code{rule} object}
\usage{
rule_init(target_outcome = "ALL", max_level = NULL, p_value = 0.05)
}
\arguments{
\item{target_outcome}{the clinical outcome that is of interest for "splitting" a cluster into two subgroups}

\item{max_level}{the maximum depth that the first clinical_outcome split can have to be of interest}

\item{p_value}{p value for association testing}
}
\value{
a \code{rule} object which can be used for tree validation
}
\description{
This function initializes a \code{rule} object to be used to check if an \code{sgi} tree is of interest
}
