% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_selection.R
\name{set_selection}
\alias{set_selection}
\title{Set selection}
\usage{
set_selection(
  rule,
  sgi_params,
  fsets,
  correction_opt = "extended_simes",
  plot = T,
  supp_plot = F,
  output_names = list(main = "set-selection-results.pdf", cluster_results =
    "set-selection-labels.xlsx", sgi_as_results =
    "set-selection-association-results.xlsx")
)
}
\arguments{
\item{rule}{A \code{rule} object containing a target clinical outcome, maximum level, and p value}

\item{sgi_params}{A \code{sgi_params} object containing information including the data, clinical outcomes, etc}

\item{fsets}{A list of lists containing the feature sets to run AutoSGI on. The names of the sets will be inferred from the names of the lists.}

\item{correction_opt}{The method used for multiple testing corrected, either simes or bonferroni}

\item{plot}{Whether to display plots of sgi trees that are of interest according to the selection rule}

\item{supp_plot}{Whether to display all the feature names of the pathway cluster in the plot file}

\item{output_names}{The output names of the AutoSGI files}
}
\value{
A vector with a boolean value for each corresponding metabolite pathway, indicating whether the sgi tree formed by that subset of metabolites is of interest
}
\description{
This is a function which uses predefined features sets of metabolites and finds \code{sgi} trees of interest
based on a \code{rule} object
}
\examples{
\dontrun{
dataset <- autosgi::qmdiab_dataset
annotations <- autosgi::qmdiab_annotations
#two categorizations of pathways
super_pathway <- annotations[, 3:3]$SUPER_PATHWAY #subsetting on metabolite rows
sub_pathway <- annotations[, 4:4]$SUB_PATHWAY
rule <- rule_init(target_outcome = "BMI", max_level = 3, p_value = 0.05) #random rule, can change based on dataset
two types of pathway annotations for this case
sup_tests <- set_selection(rule, sgi::qmdiab_plasma, sgi::qmdiab_clin, super_pathway, minsize = 18, plot = FALSE)
sub_tests <- set_selection(rule, sgi::qmdiab_plasma, sgi::qmdiab_clin, sub_pathway, minsize = 18, plot = FALSE)
print(sup_tests)
print(sub_tests)
}

}
