% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgi_params_init.R
\name{sgi_params_init}
\alias{sgi_params_init}
\title{SGI parameter intialization}
\usage{
sgi_params_init(
  dataset,
  clins,
  minsize = 18,
  user_defined_tests = c(),
  sgi_distance = "euclidean",
  sgi_linkage = "ward.D2"
)
}
\arguments{
\item{dataset}{A dataset with metabolite columns}

\item{clins}{}

\item{annotations}{A dataframe with the first column consisting on the metabolite names and the second column consisting of the corresponding annotations}

\item{method}{The method to perform reduction, ssgsea or pseudotime}
}
\value{
Always returns an updated dataframe with one column for each annotation
}
\description{
This is a function which initializes the standardized SGI parameters across all AutoSGI runs and checks for data validity
}
\examples{
\dontrun{
}


}
