% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform data}
\usage{
transform_data(dataset, annotations, method = "ssgsea")
}
\arguments{
\item{dataset}{A dataset with metabolite columns}

\item{annotations}{A dataframe with the first column consisting on the metabolite names and the second column consisting of the corresponding annotations}

\item{method}{The method to perform reduction, ssgsea or pseudotime}
}
\value{
Always returns an updated dataframe with one column for each annotation
}
\description{
This is a function which performs interpretable dimensionality reduction on the data given by ssgsea or pseudotime
}
\examples{
\dontrun{
}


}
