% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-validation.R
\name{tree_validation}
\alias{tree_validation}
\title{Validate a \code{sgi} tree based on selection rule}
\usage{
tree_validation(sgi_tree, rule)
}
\arguments{
\item{sgi_tree}{a \code{sgi} object generated from metabolite features and clinical outcomes}

\item{target_outcome}{the clinical outcome that is of interest}

\item{max_level}{the maximum depth that the first target outcome can have to be of interest}

\item{p_value}{the p value used for association testing}
}
\value{
A boolean value, with \code{TRUE} denoting a tree of interest and \code{FALSE} otherwise, all based on the selection rule
}
\description{
This is a function which determines if a tree is of interest based on a \code{rule} object
}
\examples{
\dontrun{
hc = hclust(dist(sgi::qmdiab_plasma), method = "ward.D2")
sg = sgi_init(hc, minsize = 18, outcomes = sgi::qmdiab_clin)
rule1 <- rule_init("BMI", 3, 0.05)
rule2 <- rule_init("BIM", 3, 0.05)
rule3 <- rule_init("Cholesterol", 3, 0.05)
rule4 <- rule_init("Cholesterol", 3, 0.01)
rule5 <- rule_init("Diabetes", 2, 0.05)
tree_validation(sg, rule1)
tree_validation(sg, rule2)
tree_validation(sg, rule3)
tree_validation(sg, rule4)
tree_validation(sg, rule5)
}

}
